scenario 4 - TTD for error page handling
big patch 2 sub

big patch 2 sub

eat this

+++ mvc-db-unit-integration-test/src/main/java/com/herbivore/springmvc/controller/GradebookController.java
@@ -42,6 +42,10 @@ public class GradebookController {
 
 	@PostMapping("/delete/student/{id}")
 	public String deleteStudent(@PathVariable int id) {
+		if (!studentAndGradeService.isStudentFound(id)) {
+			return "error";
+		}
+
 		studentAndGradeService.deleteStudent(id);
 		return "redirect:/";
 	}
